+++ db/migrate/20211014190537_create_materials.rb
@@ -2,8 +2,8 @@ class CreateMaterials < ActiveRecord::Migration[6.1]
   def change
     create_table :materials do |t|
       t.string :item
-      t.integer :price
-      t.integer :inventory_quantity
+      t.float :price
+      t.float :inventory_quantity
 
       t.timestamps
     end
