+++ db/migrate/20211014190554_create_invoice_materials.rb
@@ -3,8 +3,8 @@ class CreateInvoiceMaterials < ActiveRecord::Migration[6.1]
     create_table :invoice_materials do |t|
       t.belongs_to :invoice, null: false, foreign_key: true
       t.belongs_to :material, null: false, foreign_key: true
-      t.integer :quantity
-      t.integer :total
+      t.float :quantity
+      t.float :total
 
       t.timestamps
     end
